reconciles JavaScript-free locale selector with ARIA best practices
Accessible checkbox-based menus (e.g., hamburger menus[1]) tend to
rely on JavaScript to update the menu's ARIA attributes in response to
changes in state.  Here we reconcile the script-free and accessibility
requirements so that:

1. the current locale and "Choose locale" menu are announced logically;
   and

2. the menu itself is annotated to be announced by screen-reader and
   navigable by keyboard.

As a bonus, each locale is now announced *in* that locale, via the A
LANG and HREFLANG attributes.

[1]: http://www.ashleysheridan.co.uk/blog/Making+an+Accessible+Hamburger+Menu